feat: added all routes using sequelize

+++ migrations/20200503150834-create_user_table.js
@@ -5,7 +5,7 @@ module.exports = {
   up: (queryInterface, Sequelize) => {
     return queryInterface.createTable("users", {
       id: {
-        type: Sequelize.INTEGER(11),
+        type: Sequelize.INTEGER(100),
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
