+++ src/models/User.js
@@ -2,7 +2,7 @@ const Sequelize = require("sequelize");
 
 module.exports = sequelize.define("User", {
   id: {
-    type: Sequelize.INTEGER(11),
+    type: Sequelize.INTEGER(100),
     allowNull: false,
     autoIncrement: true,
     primaryKey: true,
